+++ CHANGELOG.md
@@ -6,6 +6,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This file is similar to the format suggested by [Keep a CHANGELOG](https://github.com/olivierlacan/keep-a-changelog).
 
 ## Unreleased
+- [Feature] Updated **Tile**, removing `onResultsLink` and adding `resultsLink` which helps with accessibility and user experience ([#1441](https://github.com/optimizely/oui/pull/1441))
+- [Patch] Updated **ButtonIcon**, by making `onClick` optional since a default noop prop was already there ([#1441](https://github.com/optimizely/oui/pull/1441))
 
 ## 48.7.4 - 2020-11-17
 - [Patch] Updated **DropdownBlockLinkText** with new `isItemSelected` prop to allow for black text ([#1440](https://github.com/optimizely/oui/pull/1440))